Enh/Fix: ensure cameras records have exact same frames pair + relay feed intersession error (issue-772) (#787)
* Enh/Fix: ensure cameras records exact same frames (issue-772)

+ fix: do not block on write to net_q, prevent frame drop on beginning of acquisition
+ enh: ensure both start and end frame are same in the 2 cameras
+ add detect frame drop

* Remove maintain of frame-idx in video_capture and handle in CameraBase

* Enh: relay eventual feed error before detection begin (issue-772)

+ enh/fix: missed relay segmentation and detection _complete cb to behavior algo thread
+ also relay _cb_on_set_feed_error and _cb_on_intersession_segmentation_finished to behavior algo thread
+ remove unused put_to_data_handler

* typehint + comment

* Remove use of camera frame timestamp to detect potential frame drop

* Enh: add detect frame drop with perf_counter on opencv cam (issue-772)

* Better runtime error msg

* Add check video capture open

* Fix: correct estimated_drop

* Default when_secs to math.nan to prevent TypeError

* add debug log

* fix import

* better log

+++ auto-trainer-video/src/autotrainer/video/video_record.py
@@ -30,20 +30,28 @@ class VideoRecordMode(IntEnum):
 
 @dataclass
 class VideoRecordProperties:
+
     project_info: Optional[ProjectInfo] = None
     """Information to determine file names and directories."""
+
     name: str = "camera"
     """Name used as part of video file names and image capture directory."""
+
     frame_size: Tuple[int, int] = (0, 0)
     """Expected shape of video frames.  Not required for image-only capture."""
+
     fps: int = 30
     """Expected FPS of video feed.  Not required for image-only capture."""
+
     record_mode: VideoRecordMode = VideoRecordMode.NONE
     """Continuous or triggered mode for video and image capture. NONE to disabled video recording."""
+
     video_rotate_interval: int = -1
     """Interval in seconds to rotate the video file.  0 to never rotate. Negative to disabled video recording."""
-    image_interval: int = 0
+
+    image_interval: float = 0
     """Interval in seconds to capture images.  Values <= 0 disable image capture."""
+
     queue_batch_size = 60
     """Number of frames to batch for passing between capture and record queues."""
